Add categories controller
