+++ Gemfile
@@ -59,6 +59,7 @@ group :development, :test do
   gem 'debug', platforms: %i[mri mingw x64_mingw]
   gem "database_cleaner"
   gem "rspec-rails"
+  gem 'rails-controller-testing'
 end
 
 group :development do
