--- app/controllers/category_controller.rb
@@ -1,38 +1,0 @@
-class CategoryController < ApplicationController
-    before_action :set_category, only: %i[show edit update destroy]
-    before_action :authenticate_user!
-
-    def index
-        @categories = Category.where(user: current_user)
-    end
-
-    def new
-        @category = Category.new
-    end
-
-    def edit; end
-
-    def create
-        @category = Category.new(category_params)
-        @category.user = current_user
-
-        respond_to do |format|
-            if @category.save
-                format.html { redirect_to categories_path, notice:'Category is successfully created.'}
-                format.json { render :show, status: :created, location: @category }
-            else
-              format.html { render :new, status: :unprocessable_entity }
-              format.json { render json: @category.errors, status: :unprocessable_entity }
-            end
-    end
-
-    private
-
-    def set_category
-        @category = Category.find(params[:id])
-    end
-
-    def category_params
-        params.require(:category).permit(:name, :icon)
-    end
-end
