+++ app/controllers/home_controller.rb
@@ -1,2 +1,3 @@
 class HomeController < ApplicationController
+    def home;end
 end
