+++ app/views/categories/_category.html.erb
@@ -1,0 +1,5 @@
+<div id="<%= dom_id category %>">
+  <div class="icon_container"><img src="<%= category.icon%>" class="w-24"></div>
+  <div><%= category.name%></div>
+  <div>$<%= category.total_amount%></div>
+</div>
