+++ app/views/categories/_form.html.erb
@@ -1,0 +1,26 @@
+<%= form_with(model: category) do |form| %>
+  <% if category.errors.any? %>
+    <div style="color: red">
+      <h2><%= pluralize(category.errors.count, "error") %> prohibited this category from being saved:</h2>
+
+      <ul>
+        <% category.errors.each do |error| %>
+          <li><%= error.full_message %></li>
+        <% end %>
+      </ul>
+    </div>
+  <% end %>
+  <div>
+    <%= form.label :name %> <br>
+    <%= form.text_field :name, placeholder: "Enter Category's Name" %>
+  </div>
+
+  <div>
+    <%= form.label :icon %> <br>
+    <%= form.text_field :icon, placeholder: "Enter Category's Icon" %>
+  </div>
+
+  <div>
+    <%= form.submit %>
+  </div>
+<% end %>
