+++ config/routes.rb
@@ -6,6 +6,6 @@ Rails.application.routes.draw do
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Defines the root path route ("/")
-   root "home#index"
-   resources :categories, only: [:index, :show, :new, :create, :destroy]
+  root "categories#index"
+ resources :categories, only: [:index, :new, :create, :destroy]
 end
